Moved FetchResultCode out to a common folder.
