+++ tests/core/Observer.spec.ts
@@ -4,8 +4,9 @@ import MockOperationStore from '../mocks/MockOperationStore';
 import Observer from '../../lib/core/Observer';
 import OperationProcessor from '../../lib/core/OperationProcessor';
 import { BlockchainClient } from '../../lib/core/Blockchain';
-import { CasClient, FetchResultCode } from '../../lib/core/Cas';
+import { CasClient } from '../../lib/core/Cas';
 import { ErrorCode } from '../../lib/core/util/RequestError';
+import { FetchResultCode } from '../../lib/common/FetchResultCode';
 import { MockTransactionStore } from '../mocks/MockTransactionStore';
 import { OperationStore } from '../../lib/core/OperationStore';
 import { SidetreeError } from '../../lib/core/Error';
